+++ app/models/early_career_payments/slug_sequence.rb
@@ -13,6 +13,7 @@ module EarlyCareerPayments
       "nqt-in-academic-year-after-itt",
       "supply-teacher",
       "entire-term-contract",
+      "employed-directly",
       "check-your-answers",
       "ineligible"
     ].freeze
@@ -26,6 +27,7 @@ module EarlyCareerPayments
     def slugs
       SLUGS.dup.tap do |sequence|
         sequence.delete("entire-term-contract") unless claim.eligibility.employed_as_supply_teacher?
+        sequence.delete("employed-directly") unless claim.eligibility.employed_as_supply_teacher?
         sequence.delete("ineligible") unless claim.eligibility.ineligible?
       end
     end