+++ spec/features/ineligible_early_career_payments_claims_spec.rb
@@ -5,11 +5,11 @@ RSpec.feature "Ineligible Teacher Early Career Payments claims" do
     visit landing_page_path(EarlyCareerPayments.routing_name)
     expect(page).to have_link(href: EarlyCareerPayments.feedback_url)
 
-    # Landing (start)
+    # [PAGE 00] - Landing (start)
     expect(page).to have_text(I18n.t("early_career_payments.landing_page"))
     click_on "Start Now"
 
-    # NQT in Academic Year after ITT
+    # [PAGE 01] - NQT in Academic Year after ITT
     expect(page).to have_text(I18n.t("early_career_payments.questions.nqt_in_academic_year_after_itt"))
 
     choose "No"
@@ -36,6 +36,33 @@ RSpec.feature "Ineligible Teacher Early Career Payments claims" do
 
     # [PAGE 05] - Do you have a contract to teach at the same school for an entire term or longer
     expect(page).to have_text(I18n.t("early_career_payments.questions.has_entire_term_contract"))
+
+    choose "No"
+    click_on "Continue"
+
+    expect(page).to have_text(I18n.t("early_career_payments.ineligible"))
+    expect(page).to have_link(href: EarlyCareerPayments.eligibility_page_url)
+    expect(page).to have_text("Based on the answers you have provided you are not eligible #{I18n.t("early_career_payments.claim_description")}")
+  end
+
+  # Employed as Supply Teacher by Private Agency
+  scenario "Supply Teacher employed directly by Private Agency" do
+    start_early_career_payments_claim
+
+    # [PAGE 04] - Are you currently employed as a supply teacher
+    expect(page).to have_text(I18n.t("early_career_payments.questions.employed_as_supply_teacher"))
+
+    choose "Yes"
+    click_on "Continue"
+
+    # [PAGE 05] - Do you have a contract to teach at the same school for an entire term or longer
+    expect(page).to have_text(I18n.t("early_career_payments.questions.has_entire_term_contract"))
+
+    choose "Yes"
+    click_on "Continue"
+
+    # [PAGE 06] - Are you employed directly by your school
+    expect(page).to have_text(I18n.t("early_career_payments.questions.employed_directly"))
 
     choose "No"
     click_on "Continue"